refactor(chat): extract MessageActionsHandler from CLVM
PR 4b-2 of the pre-paging cleanup, the final handler extraction. Pulls
the message-actions cluster (send / edit / delete / react / scroll-to /
mark-as-read / forward / reply / battle-dice / reaction-details) and the
sender helpers out of ConversationListViewModel.kt. With this, the VM is
~750 lines lighter than at the start of 4b-2 and ~2250 lines lighter
than at the start of the cleanup chain.

Cumulative VM trajectory:
  pre-PR-4   3564
  post-PR-4  2625  (mediaDownload + attachment handlers)
  post-4b-1  2183  (conversationLifecycle handler)
  post-4b-2  1312  (this PR)

What moved
- 20 action arms: SendMessage, EditMessage, EditMessageSave,
  CancelEditMessage, DeleteMessage, ScrollToMessageId,
  ClearHighlightedMessage, DeleteMessageForEveryone,
  DeleteMessageForMe, MarkAsRead, ToggleReaction, SendFile,
  ReplyToMessage, BattleDiceRoll, CancelBattleDiceRoll,
  ForwardMessage, ForwardMessageSelectRecipient, ForwardMessageSend,
  CancelReplyToMessage, ShowReactionDetails, HideReactionDetails.
- 6 helpers: editMessage, addMessage, replyToMessage,
  addMessageWithFiles, loadReactionDetails, processPendingSharedContent.
- pendingMessageId field (now `internal var` on the handler so
  loadMessagesForConversation can still clear it after a send lands).

Cross-handler wiring
- AttachmentHandler.handleStopRecording previously took a VM-provided
  ::addMessageWithFiles lambda. After the move, that lambda points at
  messageActionsHandler::addMessageWithFiles. This requires
  messageActionsHandler to be initialized BEFORE attachmentHandler in
  the property order; reordered accordingly.
- MessageActionsHandler.addMessageWithFiles still needs
  AttachmentHandler.ensureThumbnail to await in-flight video poster
  extraction. Solved with an `ensureThumbnail` lambda passed at ctor
  time: `{ f -> attachmentHandler.ensureThumbnail(f) }`. The lambda
  closes over `this` and resolves attachmentHandler at call time, so
  the circular dependency is broken without lateinit. Kotlin's type
  inference required explicit type annotations on both handler
  properties to resolve the recursive type check.
- VM's selectConversation now calls
  messageActionsHandler.processPendingSharedContent(conversationId)
  inside its viewModelScope.launch — same coroutine semantics.
- VM's loadMessagesForConversation now writes
  messageActionsHandler.pendingMessageId = null instead of the
  formerly-VM-private var.

Cleanup
- Dropped 22 imports that only existed for the moved code (PayloadRenderer,
  toCombinedPayloadBundle, toMessageDataType, truncateToCodePoints,
  ReplyPreview, ToggleReactionResultType, MAX_REACTIONS_PER_USER_PER_MESSAGE,
  ReactionDisplayItem, AttachmentInput, MessageAttachmentBuilder,
  LocalAttachmentContext, ImageHeaderParser, ImageUtils, convertHeicToJpeg,
  UnixTimeUtc, chat_message_forwarded, chat_reactions_limit_reached,
  DiscardDraft + ConversationListUiDialog.DeleteMessage, mimeType, name,
  resolveContentType). Caught one false positive — distinctByEmoji is still
  used by the _messagesUiState initializer — and restored it on the next
  compile.
- Removed an orphaned kdoc block left over from PR 4b-1's
  runDeleteConversationFlow deletion (it had been sitting above
  editMessage when that helper was unrelated; cleaning it up now since the
  whole helper region is being deleted anyway).

Verification
- All 6 compile targets green: homebase-chat:compileKotlinJvm /
  compileAndroidMain / compileKotlinIosSimulatorArm64,
  homebase-core:compileKotlinJvm, androidApp:compileDebugKotlin,
  desktopApp:compileKotlinJvm.
- homebase-chat:jvmTest --rerun-tasks green.
- The verbatim-move grep check passes: zero `viewModelScope` /
  `_uiState` / `_messagesUiState` references remain in the new
  MessageActionsHandler.kt body (renamed to scope/uiState/messagesUiState).

Not changed
- Public VM API: uiState, messagesUiState, onAction, eventConsumed,
  dialogClosed, conversationSearchTextState, messageInputTextState,
  messagesSearchTextState, ownerSession — all byte-identical.
- ConversationListScreen.kt, Koin module, nav graph, cross-module
  callers.
- The 23-param VM constructor.

What's left in the VM (~1312 lines): init pipelines, the 9
search/navigation arms, selectConversation /
loadMessagesForConversation / getScrollPosition / sendEvent /
updateListContent / updateMessageSearchResults helpers, and the dispatch
table. This is the natural state-coordinator shape; per the plan, no
further extraction is recommended.
